@@ -2,31 +2,31 @@ package sh.adelessfox.odradek.game.hfw.rtti;
 
 import sh.adelessfox.odradek.hashing.HashFunction;
 import sh.adelessfox.odradek.io.BinaryReader;
+import sh.adelessfox.odradek.rtti.*;
 import sh.adelessfox.odradek.rtti.data.Ref;
 import sh.adelessfox.odradek.rtti.data.Value;
 import sh.adelessfox.odradek.rtti.factory.TypeFactory;
 import sh.adelessfox.odradek.rtti.io.AbstractTypeReader;
-import sh.adelessfox.odradek.rtti.runtime.*;
 
 import java.io.IOException;
-import java.lang.reflect.Array;
 import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
 import java.util.List;
 
 import static sh.adelessfox.odradek.game.hfw.rtti.HorizonForbiddenWest.RTTIRefObject;
 
 public class HFWTypeReader extends AbstractTypeReader {
-    public record ObjectInfo(ClassTypeInfo type, RTTIRefObject object) {}
+    public record ObjectInfo(ClassTypeInfo type, RTTIRefObject object) {
+    }
 
     public static <T> T readCompound(Class<T> cls, BinaryReader reader, TypeFactory factory) throws IOException {
-        return cls.cast(new HFWTypeReader().readCompound(factory.get(cls), reader, factory));
+        var type = factory.get(cls.getSimpleName()).asClass();
+        return cls.cast(new HFWTypeReader().readCompound(type, reader, factory));
     }
 
     public ObjectInfo readObject(BinaryReader reader, TypeFactory factory) throws IOException {
         var hash = reader.readLong();
         var size = reader.readInt();
-        var type = factory.get(HFWTypeId.of(hash));
+        var type = factory.get(HFWTypeId.of(hash)).asClass();
 
         var start = reader.position();
         var object = readCompound(type, reader, factory);
@@ -44,9 +44,9 @@ public class HFWTypeReader extends AbstractTypeReader {
     }
 
     @Override
-    @SuppressWarnings("DuplicateBranchesInSwitch")
     protected Object readAtom(AtomTypeInfo info, BinaryReader reader, TypeFactory factory) throws IOException {
-        return switch (info.name().name()) {
+        var name = info.base().isPresent() ? info.base().get().name() : info.name();
+        return switch (name) {
             // Simple types
             case "bool" -> reader.readByteBoolean();
             case "wchar", "tchar" -> (char) reader.readShort();
@@ -62,27 +62,6 @@ public class HFWTypeReader extends AbstractTypeReader {
             case "String" -> readString(reader);
             case "WString" -> readWString(reader);
 
-            // Aliases
-            case "AnimationEventID" -> reader.readInt();
-            case "AnimationNodeID" -> reader.readShort();
-            case "AnimationSet" -> reader.readInt();
-            case "AnimationTagID" -> reader.readInt();
-            case "ClientID" -> reader.readByte();
-            case "EntitySoundID" -> reader.readInt();
-            case "EntitySoundParamID" -> reader.readInt();
-            case "Filename" -> readString(reader);
-            case "JointTransformList" -> reader.readLong();
-            case "LinearGainFloat" -> reader.readFloat();
-            case "MaterialType" -> reader.readShort();
-            case "MusicTime" -> reader.readLong();
-            case "PhysicsCollisionFilterInfo" -> reader.readInt();
-            case "ProgramParameterHandle" -> reader.readInt();
-            case "RelativeGainInt" -> reader.readInt();
-            case "RenderEffectFeatureSet" -> reader.readByte();
-            case "SoundGroupIndex" -> reader.readByte();
-            case "SoundVoicePluginId" -> reader.readInt();
-            case "TemplateWaveNodeEnumValue" -> reader.readInt();
-
             default -> throw new IllegalArgumentException("Unknown atom type: " + info.name());
         };
     }
@@ -95,7 +74,7 @@ public class HFWTypeReader extends AbstractTypeReader {
             case Integer.BYTES -> reader.readInt();
             default -> throw new IllegalArgumentException("Unexpected enum size: " + info.size());
         };
-        if (info.isSet()) {
+        if (info instanceof EnumSetTypeInfo) {
             return info.setOf(value);
         } else {
             return info.valueOf(value);
@@ -104,7 +83,7 @@ public class HFWTypeReader extends AbstractTypeReader {
 
     @Override
     protected Object readContainer(ContainerTypeInfo info, BinaryReader reader, TypeFactory factory) throws IOException {
-        return switch (info.name().name()) {
+        return switch (info.containerType()) {
             case "HashMap", "HashSet" -> readHashContainer(info, reader, factory);
             default -> readSimpleContainer(info, reader, factory);
         };
@@ -116,53 +95,74 @@ public class HFWTypeReader extends AbstractTypeReader {
     }
 
     private Object readSimpleContainer(ContainerTypeInfo info, BinaryReader reader, TypeFactory factory) throws IOException {
-        var itemInfo = info.itemType().get();
-        var itemType = itemInfo.type();
         var count = reader.readInt();
+        var item = info.itemType();
 
         // Fast path
-        if (itemType == byte.class) {
-            return reader.readBytes(count);
-        } else if (itemType == short.class) {
-            return reader.readShorts(count);
-        } else if (itemType == int.class) {
-            return reader.readInts(count);
-        } else if (itemType == long.class) {
-            return reader.readLongs(count);
+        if (item instanceof AtomTypeInfo atom) {
+            var name = atom.base().isPresent() ? atom.base().get().name() : atom.name();
+            switch (name) {
+                case "bool" -> {
+                    var items = new boolean[count];
+                    for (int i = 0; i < count; i++) {
+                        items[i] = reader.readByteBoolean();
+                    }
+                    return items;
+                }
+                case "wchar", "tchar" -> {
+                    var items = new char[count];
+                    for (int i = 0; i < count; i++) {
+                        items[i] = (char) reader.readShort();
+                    }
+                    return items;
+                }
+                case "uint8", "int8" -> {
+                    return reader.readBytes(count);
+                }
+                case "uint16", "int16" -> {
+                    return reader.readShorts(count);
+                }
+                case "uint", "int", "uint32", "int32", "ucs4" -> {
+                    return reader.readInts(count);
+                }
+                case "uint64", "int64" -> {
+                    return reader.readLongs(count);
+                }
+                case "HalfFloat" -> {
+                    return reader.readHalfs(count);
+                }
+                case "float" -> {
+                    return reader.readFloats(count);
+                }
+                case "double" -> {
+                    return reader.readDoubles(count);
+                }
+            }
         }
 
-        var array = Array.newInstance(itemType, count);
+        var items = new Object[count];
         for (int i = 0; i < count; i++) {
-            Array.set(array, i, read(itemInfo, reader, factory));
+            items[i] = read(item, reader, factory);
         }
 
-        if (info.type() == List.class) {
-            return Arrays.asList((Object[]) array);
-        } else {
-            return array;
-        }
+        return List.of(items);
     }
 
     private Object readHashContainer(ContainerTypeInfo info, BinaryReader reader, TypeFactory factory) throws IOException {
-        var itemInfo = info.itemType().get();
-        var itemType = itemInfo.type();
         var count = reader.readInt();
+        var item = info.itemType();
 
-        var array = Array.newInstance(itemType, count);
+        var items = new Object[count];
         for (int i = 0; i < count; i++) {
             // NOTE: Hash is based on the key - for HashMap, and on the value - for HashSet
             //       We don't actually need to store or use it - but we'll have to compute it
             //       when serialization support is added
             int hash = reader.readInt();
-            Array.set(array, i, read(itemInfo, reader, factory));
+            items[i] = read(item, reader, factory);
         }
 
         // TODO: Use specialized type (Map, Set, etc.)
-        if (info.type() == List.class) {
-            return Arrays.asList((Object[]) array);
-        } else {
-            return array;
-        }
+        return List.of(items);
     }
 
     private static String readString(BinaryReader reader) throws IOException {
