@@ -2,7 +2,7 @@ package sh.adelessfox.odradek.game.hfw.ui.renderers;
 
 import sh.adelessfox.odradek.game.hfw.game.ForbiddenWestGame;
 import sh.adelessfox.odradek.game.hfw.rtti.HorizonForbiddenWest.GGUUID;
-import sh.adelessfox.odradek.rtti.runtime.TypeInfo;
+import sh.adelessfox.odradek.rtti.TypeInfo;
 import sh.adelessfox.odradek.ui.Renderer;
 
 import java.util.Optional;
