@@ -1,8 +1,8 @@
 package sh.adelessfox.odradek.game.hfw.ui.renderers;
 
 import sh.adelessfox.odradek.game.Game;
-import sh.adelessfox.odradek.rtti.runtime.AtomTypeInfo;
-import sh.adelessfox.odradek.rtti.runtime.TypeInfo;
+import sh.adelessfox.odradek.rtti.AtomTypeInfo;
+import sh.adelessfox.odradek.rtti.TypeInfo;
 import sh.adelessfox.odradek.ui.Renderer;
 
 import java.util.Optional;
@@ -22,6 +22,6 @@ public class UnsignedIntRenderer implements Renderer<Number, Game> {
 
     @Override
     public boolean supports(TypeInfo info) {
-        return info instanceof AtomTypeInfo && info.name().fullName().startsWith("uint");
+        return info instanceof AtomTypeInfo && info.name().startsWith("uint");
     }
 }
