@@ -1,15 +1,17 @@
 package sh.adelessfox.odradek.game.hfw.rtti;
 
 import sh.adelessfox.odradek.hashing.HashFunction;
+import sh.adelessfox.odradek.rtti.TypeInfo;
 import sh.adelessfox.odradek.rtti.factory.AbstractTypeFactory;
 import sh.adelessfox.odradek.rtti.factory.TypeId;
-import sh.adelessfox.odradek.rtti.factory.TypeName;
-import sh.adelessfox.odradek.rtti.runtime.TypeInfo;
 
 import java.lang.invoke.MethodHandles;
+import java.math.BigInteger;
+import java.net.URL;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Map;
 
 public class HFWTypeFactory extends AbstractTypeFactory {
     public HFWTypeFactory() {
@@ -18,29 +20,56 @@ public class HFWTypeFactory extends AbstractTypeFactory {
 
     @Override
     protected TypeId computeTypeId(TypeInfo info) {
-        var name = "00000001_" + getInternalName(info.name());
+        var name = "00000001_" + info.name();
         var hash = HashFunction.murmur3().hash(name).asLong();
         return HFWTypeId.of(hash);
     }
 
     @Override
-    protected void sortSerializableAttrs(List<OrderedAttr> attrs) {
+    protected void sortOrderedAttributes(List<OrderedAttr> attrs) {
         quicksort(attrs, Comparator.comparingInt(OrderedAttr::offset), 0, attrs.size() - 1, 0);
     }
 
     @Override
-    protected void filterSerializableAttrs(List<OrderedAttr> attrs) {
-        // Remove save state attribute
-        attrs.removeIf(attr -> (attr.info().flags() & 2) != 0);
-        // Remove non-"serializable" attributes. They include holders for MsgReadBinary data
-        attrs.removeIf(attr -> !attr.serializable());
+    protected void filterOrderedAttributes(List<OrderedAttr> attrs) {
+        // Remove save-state attribute
+        attrs.removeIf(attr -> (attr.attr().flags() & 2) != 0);
     }
 
-    private static String getInternalName(TypeName name) {
-        return switch (name) {
-            case TypeName.Simple(var n) -> n;
-            case TypeName.Parameterized(var n, var a) -> n + '_' + getInternalName(a);
-        };
+    @Override
+    protected URL getTypes() {
+        return getClass().getResource("/types.json");
+    }
+
+    @Override
+    protected URL getExtensions() {
+        return getClass().getResource("/extensions.json");
+    }
+
+    @Override
+    protected Map<String, Class<?>> getBuiltins() {
+        return Map.ofEntries(
+            Map.entry("HalfFloat", float.class),
+            Map.entry("float", float.class),
+            Map.entry("double", double.class),
+            Map.entry("bool", boolean.class),
+            Map.entry("wchar", char.class),
+            Map.entry("ucs4", int.class),
+            Map.entry("uint", int.class),
+            Map.entry("uint8", byte.class),
+            Map.entry("uint16", short.class),
+            Map.entry("uint32", int.class),
+            Map.entry("uint64", long.class),
+            Map.entry("uint128", BigInteger.class),
+            Map.entry("uintptr", long.class),
+            Map.entry("int", int.class),
+            Map.entry("int8", byte.class),
+            Map.entry("int16", short.class),
+            Map.entry("int32", int.class),
+            Map.entry("int64", long.class),
+            Map.entry("String", String.class),
+            Map.entry("WString", String.class)
+        );
     }
 
     private static <T> int quicksort(List<T> items, Comparator<T> comparator, int left, int right, int state) {
