@@ -1,47 +1,15 @@
 package sh.adelessfox.odradek.game.hfw.rtti.callbacks;
 
+import sh.adelessfox.odradek.game.hfw.rtti.HorizonForbiddenWest.DataBufferResource;
 import sh.adelessfox.odradek.io.BinaryReader;
-import sh.adelessfox.odradek.rtti.Attr;
 import sh.adelessfox.odradek.rtti.data.ExtraBinaryDataCallback;
 import sh.adelessfox.odradek.rtti.factory.TypeFactory;
 
 import java.io.IOException;
 
-public class DataBufferResourceCallback implements ExtraBinaryDataCallback<DataBufferResourceCallback.DataBuffer> {
-    public interface DataBuffer {
-        @Attr(name = "Count", type = "uint32", position = 0)
-        int count();
-
-        void count(int value);
-
-        @Attr(name = "Streaming", type = "bool", position = 1)
-        boolean streaming();
-
-        void streaming(boolean value);
-
-        @Attr(name = "Flags", type = "uint32", position = 2)
-        int flags();
-
-        void flags(int value);
-
-        @Attr(name = "Format", type = "uint32", position = 3)
-        int format();
-
-        void format(int value);
-
-        @Attr(name = "Stride", type = "uint32", position = 4)
-        int stride();
-
-        void stride(int value);
-
-        @Attr(name = "Data", type = "Array<uint8>", position = 5)
-        byte[] data();
-
-        void data(byte[] value);
-    }
-
+public class DataBufferResourceCallback implements ExtraBinaryDataCallback<DataBufferResource> {
     @Override
-    public void deserialize(BinaryReader reader, TypeFactory factory, DataBuffer object) throws IOException {
+    public void deserialize(BinaryReader reader, TypeFactory factory, DataBufferResource object) throws IOException {
         var count = reader.readInt();
         if (count == 0) {
             return;
