@@ -1,20 +1,16 @@
 package sh.adelessfox.odradek.game.hfw.rtti.data;
 
 import sh.adelessfox.odradek.game.hfw.rtti.HorizonForbiddenWest.StreamingDataSource;
-import sh.adelessfox.odradek.rtti.Attr;
 
 public interface StreamingDataSourceExtension {
-    @Attr(name = "Locator", type = "uint64", position = 0, offset = 0)
-    long locator();
-
-    void locator(long locator);
-
     default int fileId() {
-        return (int) (locator() & 0xffffff);
+        var dataSource = (StreamingDataSource) this;
+        return (int) (dataSource.locator() & 0xffffff);
     }
 
     default int fileOffset() {
-        return (int) (locator() >>> 24);
+        var dataSource = (StreamingDataSource) this;
+        return (int) (dataSource.locator() >>> 24);
     }
 
     default boolean isValid() {
