@@ -1,111 +1,70 @@
 package sh.adelessfox.odradek.rtti.generator;
 
-import com.palantir.javapoet.AnnotationSpec;
-import com.palantir.javapoet.JavaFile;
-import com.palantir.javapoet.TypeName;
-import com.palantir.javapoet.TypeSpec;
-import sh.adelessfox.odradek.rtti.generator.data.ClassTypeInfo;
-import sh.adelessfox.odradek.rtti.generator.data.TypeInfo;
+import sh.adelessfox.odradek.rtti.runtime.TypeContext;
 
 import javax.annotation.processing.AbstractProcessor;
-import javax.annotation.processing.Generated;
 import javax.annotation.processing.RoundEnvironment;
 import javax.lang.model.SourceVersion;
 import javax.lang.model.element.Element;
-import javax.lang.model.element.Modifier;
 import javax.lang.model.element.ModuleElement;
 import javax.lang.model.element.TypeElement;
-import javax.lang.model.type.DeclaredType;
 import javax.lang.model.type.MirroredTypeException;
 import javax.lang.model.type.TypeMirror;
 import javax.tools.StandardLocation;
 import java.io.IOException;
-import java.io.Writer;
-import java.lang.annotation.Annotation;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.Objects;
+import java.lang.constant.ClassDesc;
+import java.net.URL;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
+import java.util.function.Supplier;
 
 public class GenerateBindingsProcessor extends AbstractProcessor {
-    private static final AnnotationSpec suppressWarningsAnnotation = AnnotationSpec.builder(SuppressWarnings.class)
-        .addMember("value", "$S", "all")
-        .build();
-    private static final AnnotationSpec generatedBindingsAnnotation = AnnotationSpec.builder(Generated.class)
-        .addMember("value", "$S", GenerateBindingsProcessor.class.getName())
-        .build();
-
     @Override
     public boolean process(Set<? extends TypeElement> annotations, RoundEnvironment roundEnv) {
+        var filer = processingEnv.getFiler();
         var messager = processingEnv.getMessager();
-        var types = processingEnv.getTypeUtils();
 
         for (Element element : roundEnv.getElementsAnnotatedWith(GenerateBindings.class)) {
             var module = (ModuleElement) element;
             var annotation = module.getAnnotation(GenerateBindings.class);
 
-            var targetName = annotation.target();
-            var className = targetName.substring(targetName.lastIndexOf('.') + 1);
-            var packageName = targetName.substring(0, targetName.lastIndexOf('.'));
-            var qualifiedName = module.getQualifiedName() + "/" + targetName;
+            var context = new TypeContext();
+            try {
+                context.load(getResourceUrl(annotation.input().types()), getResourceUrl(annotation.input().extensions()));
+            } catch (IOException e) {
+                messager.printError("An error occurred while reading type definitions: " + e.getMessage(), module);
+            }
 
-            messager.printNote("generating type bindings " + targetName + " from " + annotation.source());
+            var targetDesc = ClassDesc.of(annotation.target());
+            var targetPackage = targetDesc.packageName();
+            var targetClass = targetDesc.displayName();
+
+            var generator = new TypeSourceGenerator(targetPackage, targetClass);
+            for (GenerateBindings.Builtin builtin : annotation.builtins()) {
+                generator.addBuiltin(builtin.type(), asTypeMirror(builtin::repr));
+            }
+            for (GenerateBindings.Callback callback : annotation.callbacks()) {
+                var type = asTypeMirror(callback::handler);
+                generator.addCallback(callback.type(), type);
+            }
+            for (GenerateBindings.Extension extension : annotation.extensions()) {
+                var type = asTypeMirror(extension::extension);
+                generator.addExtension(extension.type(), type);
+            }
+
+            generator.addAll(context.getAll().toList());
 
             try {
-                var context = new TypeContext();
-                try (var reader = Files.newBufferedReader(getModuleRoot().resolve(annotation.source()))) {
-                    context.load(reader);
-                }
-
-                // Report missing callbacks right away. May be replaced with errors later on
-                reportMissingCallbacks(context, annotation);
-
-                var generator = new TypeGenerator();
-                for (GenerateBindings.Callback callback : annotation.callbacks()) {
-                    // Shenanigans to access <T> of ExtraBinaryDataCallback
-                    var handlerType = (TypeElement) types.asElement(getAnnotationValueMirror(callback, GenerateBindings.Callback::handler));
-                    var handlerParent = (DeclaredType) handlerType.getInterfaces().getFirst();
-                    var holderType = types.asElement(handlerParent.getTypeArguments().getFirst());
-                    generator.addCallback(callback.type(), TypeName.get(handlerType.asType()));
-                    generator.addExtension(callback.type(), TypeName.get(holderType.asType()));
-                }
-
-                for (GenerateBindings.Builtin builtin : annotation.builtins()) {
-                    var javaType = getAnnotationValueMirror(builtin, GenerateBindings.Builtin::javaType);
-                    generator.addBuiltin(builtin.type(), TypeName.get(javaType));
-                }
-
-                for (GenerateBindings.Extension extension : annotation.extensions()) {
-                    var extensionType = getAnnotationValueMirror(extension, GenerateBindings.Extension::extension);
-                    generator.addExtension(extension.type(), TypeName.get(extensionType));
-                }
-
-                var builder = TypeSpec.interfaceBuilder(className)
-                    .addModifiers(Modifier.PUBLIC)
-                    .addAnnotation(suppressWarningsAnnotation)
-                    .addAnnotation(generatedBindingsAnnotation);
-
-                for (TypeInfo info : context.types()) {
-                    if (skipType(info)) {
-                        continue;
-                    }
-                    generator.generate(info).ifPresent(builder::addType);
-                }
-
-                try (Writer writer = processingEnv.getFiler().createSourceFile(qualifiedName, element).openWriter()) {
-                    JavaFile.builder(packageName, builder.build())
-                        .indent("\t")
-                        .addFileComment("This file was autogenerated. Do not edit!")
-                        .build().writeTo(writer);
-                }
-            } catch (Exception e) {
-                messager.printError("failed to generate: " + e);
+                generator.build().writeTo(filer);
+            } catch (IOException e) {
+                messager.printError("Am error occurred while writing generated code: " + e.getMessage());
             }
         }
 
         return true;
+    }
+
+    private URL getResourceUrl(String name) throws IOException {
+        return processingEnv.getFiler().getResource(StandardLocation.CLASS_OUTPUT, "", name).toUri().toURL();
     }
 
     @Override
@@ -118,44 +77,12 @@ public class GenerateBindingsProcessor extends AbstractProcessor {
         return SourceVersion.latestSupported();
     }
 
-    private static boolean skipType(TypeInfo info) {
-        return info instanceof ClassTypeInfo cls && cls.isAssignableTo("ExportedSymbolGroup");
-    }
-
-    private void reportMissingCallbacks(TypeContext context, GenerateBindings annotation) {
-        Set<String> typesMissingCallback = context.types().stream()
-            .filter(info -> info instanceof ClassTypeInfo cls && cls.messages().contains("MsgReadBinary"))
-            .map(info -> info.typeName().fullName())
-            .collect(Collectors.toSet());
-
-        for (GenerateBindings.Callback callback : annotation.callbacks()) {
-            typesMissingCallback.remove(callback.type());
-        }
-
-        for (String type : typesMissingCallback) {
-            processingEnv.getMessager().printWarning("no extra binary data callback for " + annotation.target() + "." + type);
-        }
-    }
-
-    private Path getModuleRoot() throws IOException {
-        var file = processingEnv.getFiler().getResource(StandardLocation.CLASS_OUTPUT, "", "dummy");
-        var path = Path.of(file.toUri());
-        while (Files.notExists(path.resolve("pom.xml"))) {
-            path = path.getParent();
-        }
-        return path;
-    }
-
-    private static <A extends Annotation> TypeMirror getAnnotationValueMirror(
-        A annotation,
-        Function<A, Class<?>> extractor
-    ) {
+    private TypeMirror asTypeMirror(Supplier<Class<?>> supplier) {
         try {
-            Class<?> ignored = extractor.apply(annotation);
-            // ^ this will always fail
+            supplier.get();
             throw new AssertionError();
         } catch (MirroredTypeException e) {
-            return Objects.requireNonNull(e.getTypeMirror());
+            return e.getTypeMirror();
         }
     }
 }
