@@ -1,6 +1,9 @@
 module odradek.rtti {
     requires odradek.core;
-    requires org.objectweb.asm;
+
+    requires com.google.gson;
+    requires java.compiler; // Required for javax.lang.model.SourceVersion
+    requires org.slf4j;
 
     exports sh.adelessfox.odradek.rtti.data;
     exports sh.adelessfox.odradek.rtti.factory;
